\name{uclidean-package}
\alias{uclidean-package}
\alias{uclidean}
\docType{package}
\title{
\packageTitle{uclidean}
}
\description{
\packageDescription{uclidean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{uclidean}
\packageIndices{uclidean}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{uclidean}

Maintainer: \packageMaintainer{uclidean}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
